+++ tests/forge/expected/codec_zenoh_frame_test.c.h
@@ -55,7 +55,10 @@ static inline int test_vector_codec_zenoh_frame(void) {
                     fprintf(stderr,
                         "FAIL: codec_zenoh_frame test_vector @SCXML L38: "
                         "field `payload` expected empty bytes, got len=%zu\n",
-                        decoded.payload_len);
+                        /* The length companion is `size_t` for a tail field but
+                         * the (possibly narrower) wire length-field for a
+                         * length-ref payload — cast so `%zu` is always valid. */
+                        (size_t)decoded.payload_len);
                     ++failures;
                 }
             }
